made changes to save variables for inhdlogin method

+++ controllers/libraries_controller.php
@@ -412,15 +412,17 @@ Class LibrariesController extends AppController
                                                     $this->data['Library']['library_status'] = 'inactive';
                                                 }
                                                 if($this->Library->save($this->data['Library'])) {
-														if($this->data['Library']['library_authentication_method'] == 'innovative_var_wo_pin' || $this->data['Library']['library_authentication_method'] == 'sip2_var' || $this->data['Library']['library_authentication_method'] == 'sip2_var_wo_pin' || $this->data['Library']['library_authentication_method'] == 'innovative_https' || $this->data['Library']['library_authentication_method'] == 'innovative_var' || $this->data['Library']['library_authentication_method'] == 'innovative_var_https' || $this->data['Library']['library_authentication_method'] == 'innovative_var_https_wo_pin' || $this->data['Library']['library_authentication_method'] == 'innovative_var_name'){
-															foreach($this->data['Variable'] as $k=>$v){
-																if($this->data['Variable'][$k]['authentication_variable'] !='' && $this->data['Variable'][$k]['authentication_response'] != '' && $this->data['Variable'][$k]['error_msg'] != ''){
-																	$data[$k] = $v;
-																	$data[$k]['library_id'] = $this->Library->id;
+														if(count($this->data['Variable']) > 0){
+															if($this->data['Library']['library_authentication_method'] == 'innovative_var_wo_pin' || $this->data['Library']['library_authentication_method'] == 'sip2_var' || $this->data['Library']['library_authentication_method'] == 'sip2_var_wo_pin' || $this->data['Library']['library_authentication_method'] == 'innovative_https' || $this->data['Library']['library_authentication_method'] == 'innovative_var' || $this->data['Library']['library_authentication_method'] == 'innovative_var_https' || $this->data['Library']['library_authentication_method'] == 'innovative_var_https_wo_pin' || $this->data['Library']['library_authentication_method'] == 'innovative_var_name'){
+																foreach($this->data['Variable'] as $k=>$v){
+																	if($this->data['Variable'][$k]['authentication_variable'] !='' && $this->data['Variable'][$k]['authentication_response'] != '' && $this->data['Variable'][$k]['error_msg'] != ''){
+																		$data[$k] = $v;
+																		$data[$k]['library_id'] = $this->Library->id;
+																	}
 																}
+																$this->Variable->deleteAll(array('library_id' => $this->Library->id));
+																$this->Variable->saveAll($data);
 															}
-															$this->Variable->deleteAll(array('library_id' => $this->Library->id));
-															$this->Variable->saveAll($data);
 														}
 														if($this->data['Libraryurl'][0]['domain_name']){
 															if($this->data['Library']['library_authentication_method'] != 'referral_url' || $this->data['Library']['library_authentication_method'] != 'user_account'){
